FW-4627 Refactor tests and fix duplication
